Fixes
Fixed the form checks so that cities cannot have numbers and added the drop down for states. Also, changed the styling for the password and the names. Email must have an at symbol and period. All forms are now required and will be directed to the missing input. Added in the about us page changes. Other CSS tweaks that make it look nice.