Branch lloyd (#2)
* Added in signup page-need further formatting

* Some more Formatting and DataBase SQL file

* Updated the links

* Button Name Change

* New Form

Added a form on the joinh us page and made some minor adjustments and fixes to About Us page.

* added in php - in progress

* Menu colors

Color changes in the slide out menu based on which page you are on.

* File name change to php

* form action

* Jump to center

* Username duplicates checked

* Form Completion

Use regular expressions to check the validity of the form and also added a feature where the form does not reset if you mess up filling it out.

* Fixes

Fixed the form checks so that cities cannot have numbers and added the drop down for states. Also, changed the styling for the password and the names. Email must have an at symbol and period. All forms are now required and will be directed to the missing input. Added in the about us page changes. Other CSS tweaks that make it look nice.

* Photos

Added some stock photos to the homepage

* commented out "agreement"

* About Us Fixes

* Test

Made a slight change

* Undid the change

* Added PayPal Clickable Icon

* fixed zipcode check and changed button

* Member's Only Pages, Limited Functionality

* Minor fixes in new pages

* My Requests Page,

If you don't like the diff formatting, easy way to switch back to normal format noted in code's comments

* Disabled buttons for Requests in Progress

* Minor changes

* added paypal files

* Reciept page for joining Stork

* Does it actually send mail though?

* paypal layout

* Paypal - in progress

* minor changes

* Set up Paypal Transaction

Balance Shows up on both Buyer and seller account

* formatting

* removed shipping

* amount

* Commented out the links to user only pages, fixed joinUs menu nav

* My Stork Quests

* payment amount confirmation

* Milestone 3